requirements.txt
tensorflow==2.19.0
numpy==1.26.0
pandas==2.2.3
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.1.post1
opencv-python==4.10.0.84
streamlit
pillow